// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const AdminRole = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  ADMIN: 'ADMIN'
} as const

export type AdminRole = (typeof AdminRole)[keyof typeof AdminRole]


export const YachtCondition = {
  NEW: 'NEW',
  USED: 'USED'
} as const

export type YachtCondition = (typeof YachtCondition)[keyof typeof YachtCondition]


export const YachtStatus = {
  DRAFT: 'DRAFT',
  AVAILABLE: 'AVAILABLE',
  SOLD: 'SOLD'
} as const

export type YachtStatus = (typeof YachtStatus)[keyof typeof YachtStatus]